Update Practical_1.Rmd
Loading Excel files

There are several options:
•	library(xlsx) with read.xlsx()
Loading the library will not work on Windows R 4.2.0. Users with this version have to install a patched version of R.
For the problem discussion see https://community.rstudio.com/t/r-session-aborted-when-loading-xlsx-package/135572/3
A patched version of R is available at: https://cran.r-project.org/bin/windows/base/rpatched.html
After installation of the patch, the problem is no longer present.
•	library (readxl) with read_excel()
Does not work for me (loading the file takes forever).
•	library(openxlsx) with read.xlsx()
Works well for me. My suggestion is officially going with openxlx but informing the students about the alternatives. I adjusted the code accordingly.

2. **Randomly sample 1000 of the blogs into a new variable and name it `blog_gender_1000`. Hint: use the `sample_n` function.**

Using dplyr::sample_n() works well. As a note, this function has been superseded in favor of slice_sample(). While sample_n() will not be deprecated in the near future, its retirement means that the developers will only perform critical bug fixes, so they recommend moving to the newer alternative. In essence, slice_sample() it does the same thing. I added the code for the new function (commented) in case you think we should include it. I did not adjust the question as the current function works fine. Let me know if you want to replace it with the newer function.

3. **Use the `head`, `tail`, and `View` functions to check the dataset. What are the differences between these functions?**

Problem: using head() and tail() creates a lot of empty space in the console (also in a later example). It is difficult to get an overview of the data frame. Solution: converting the data frame to a tibble using as_tibble(). The new output when using head() and tail() is significantly easier to read. There are no problems with tibbles in the ongoing code. I think this is more elegant, so I implemented it in this question.

Optional: Could ask participants to extract the text and gender of the 135th entry (could be nice as it shows the entire text and how to index tibbles). The answer would be:
blog_gender_1000$blog[135]	and
blog_gender_1000$gender[135]
I did not add this optional part. If you like it I can implement it as a short additional question.

14. **Except `computer_531`, `imdb_500`, and `blog_gender_1000` remove the other files from the RStudio environment**

I add gc() # "garbage collector" to free up memory of removed variables

19. **Now we want to process the `computer_531` data and separate aspects and sentiments for each record. First, use regular expression to extract the characters at the beginning of each line until `##`. Apply this only for the first 20 reviews in data. Use `str_extract` or `gsub` function.**

Note: for this to work well computer_531 has to be a data frame, not a tibble. I leave it as a data frame.